+++ docs/src/changelog.md
@@ -10,6 +10,12 @@ If your currently installed *Quality-time* version is not the penultimate versio
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the release/release.py script with the new release version and release date. -->
 
+## [Unreleased]
+
+### Added
+
+- Allow for ignoring draft merge requests when using GitLab. Closes [#12327](https://github.com/ICTU/quality-time/issues/12327).
+
 ## v5.48.5 - 2026-02-27
 
 ### Fixed